rename zip-entry.ts => entry.ts
